break up the thick controller and remove some unused errors
