+++ services/platform/server/system/controller.go
@@ -21,13 +21,10 @@ import (
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/go-git/go-git/v5/storage/memory"
-	"github.com/google/uuid"
-	"github.com/robfig/cron/v3"
 	kvv1 "github.com/steady-bytes/draft/api/core/registry/key_value/v1"
 	"github.com/steady-bytes/draft/pkg/chassis"
 	"golang.org/x/mod/semver"
 	"golang.org/x/sync/errgroup"
-	"golang.zx2c4.com/wireguard/wgctrl/wgtypes"
 )
 
 type (
@@ -36,66 +33,6 @@ type (
 		OS
 		Containers
 		Device
-	}
-	Daemon interface {
-		// ShutdownHost will shutdown the host machine running Home Cloud.
-		ShutdownHost() error
-		// RestartHost will restart the host machine running Home Cloud
-		RestartHost() error
-		// ChangeDaemonVersion will update the NixOS config with a new Daemon version
-		// and switch to it.
-		ChangeDaemonVersion(cmd *dv1.ChangeDaemonVersionCommand) error
-		// AddMdnsHost adds a host to the avahi mDNS server managed by the daemon
-		AddMdnsHost(hostname string) error
-		// RemoveMdnsHost removes a host to the avahi mDNS server managed by the daemon
-		RemoveMdnsHost(hostname string) error
-		// UploadFileStream will stream a file in chunks as an upload to the daemon
-		UploadFileStream(ctx context.Context, logger chassis.Logger, buf io.Reader, fileId, fileName string) (string, error)
-	}
-	OS interface {
-		// CheckForOSUpdates will run the Nix commands to check for any NixOS updates to install.
-		CheckForOSUpdates(ctx context.Context, logger chassis.Logger) (*v1.CheckForSystemUpdatesResponse, error)
-		// InstallOSUpdate will install the update (if any) generated by CheckForOSUpdates.
-		InstallOSUpdate() error
-		// AutoUpdateOS will check for and install any OS (including Daemon) updates on a schedule. It is
-		// designed to be called at bootup.
-		AutoUpdateOS(logger chassis.Logger)
-		// UpdateOS will check for and install any OS (including Daemon) updates one time.
-		UpdateOS(ctx context.Context, logger chassis.Logger) error
-		// EnableWireguard will initialize the Wireguard server on the host and save the configuration to Blueprint
-		EnableWireguard(ctx context.Context, logger chassis.Logger) error
-		// DisableWireguard will disable the Wireguard server on the host and delete the configuration from Blueprint
-		DisableWireguard(ctx context.Context, logger chassis.Logger) error
-	}
-	Containers interface {
-		// SetSystemImage will update the image for a system container.
-		SetSystemImage(cmd *dv1.SetSystemImageCommand) error
-		// CheckForContainerUpdates will compare current system container images against the latest ones
-		// available and return the result.
-		CheckForContainerUpdates(ctx context.Context, logger chassis.Logger) ([]*v1.ImageVersion, error)
-		// AutoUpdateContainers will check for and install any container updates on a schedule. It is
-		// designed to be called at bootup.
-		AutoUpdateContainers(logger chassis.Logger)
-		// UpdateContainers will check for and install any container updates one time.
-		UpdateContainers(ctx context.Context, logger chassis.Logger) error
-	}
-	Device interface {
-		// GetServerSettings returns the current server settings after filtering out the
-		// admin username and password.
-		GetServerSettings(ctx context.Context) (*v1.DeviceSettings, error)
-		// SetServerSettings updates the settings on the server with the given values
-		SetServerSettings(ctx context.Context, logger chassis.Logger, settings *v1.DeviceSettings) error
-		// IsDeviceSetup checks if the device is already setup by checking if the DEFAULT_DEVICE_SETTINGS_KEY key exists in the key-value store
-		// with the default settings model
-		IsDeviceSetup(ctx context.Context) (bool, error)
-		// InitializeDevice initializes the device with the given settings. It first checks if the device is already setup
-		// Uses the user-provided password to set the password for the "admin" user on the device
-		// and save the remaining settings in the key-value store
-		InitializeDevice(ctx context.Context, logger chassis.Logger, settings *v1.DeviceSettings) error
-		// Login receives a username and password and returns a token.
-		//
-		// NOTE: the token is unimplemented so this only checks the validity of the password for now
-		Login(ctx context.Context, username, password string) (string, error)
 	}
 
 	controller struct {
@@ -121,14 +58,7 @@ func NewController(logger chassis.Logger, broadcaster async.Broadcaster) Control
 }
 
 const (
-	ErrDeviceAlreadySetup          = "device already setup"
-	ErrFailedToCreateSettings      = "failed to create settings"
-	ErrFailedToSaveSettings        = "failed to save settings"
-	ErrFailedToGetSettings         = "failed to get settings"
-	ErrFailedToSetSettings         = "failed to save settings"
 	ErrFailedToBuildSeedGetRequest = "failed to build get request for seed"
-	ErrFailedToGetSeedValue        = "failed to get seed value"
-	ErrFailedToUnmarshalSeedValue  = "failed to unmarshal seed value"
 
 	homeCloudCoreRepo                = "https://github.com/home-cloud-io/core"
 	homeCloudCoreTrunk               = "main"
@@ -140,635 +70,6 @@ const (
 	// will probably want to support multiple interfaces (e.g. one for trusted mobile clients and another for federated servers)
 	DefaultWireguardInterface = "wg0"
 )
-
-// DAEMON
-
-func (c *controller) ShutdownHost() error {
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_Shutdown{},
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *controller) RestartHost() error {
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_Restart{},
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *controller) ChangeDaemonVersion(cmd *dv1.ChangeDaemonVersionCommand) error {
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_ChangeDaemonVersionCommand{
-			ChangeDaemonVersionCommand: cmd,
-		},
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *controller) InstallOSUpdate() error {
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_InstallOsUpdateCommand{},
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *controller) SetSystemImage(cmd *dv1.SetSystemImageCommand) error {
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_SetSystemImageCommand{
-			SetSystemImageCommand: cmd,
-		},
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *controller) AddMdnsHost(hostname string) error {
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_AddMdnsHostCommand{
-			AddMdnsHostCommand: &dv1.AddMdnsHostCommand{
-				Hostname: hostname,
-			},
-		},
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *controller) RemoveMdnsHost(hostname string) error {
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_RemoveMdnsHostCommand{
-			RemoveMdnsHostCommand: &dv1.RemoveMdnsHostCommand{
-				Hostname: hostname,
-			},
-		},
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *controller) UploadFileStream(ctx context.Context, logger chassis.Logger, buf io.Reader, fileId, fileName string) (string, error) {
-	logger.Info("uploading file")
-	var listenerErr error
-	done := make(chan bool)
-	go func() {
-		listenerErr = async.RegisterListener(ctx, c.broadcaster, &async.ListenerOptions[*dv1.UploadFileReady]{
-			Callback: func(event *dv1.UploadFileReady) (bool, error) {
-				if event.Id == fileId {
-					done <- true
-					return true, nil
-				}
-				return false, nil
-			},
-			Timeout: 5 * time.Second,
-		}).Listen(ctx)
-		if listenerErr != nil {
-			done <- true
-		}
-	}()
-
-	// prepare upload to daemon
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_UploadFileRequest{
-			UploadFileRequest: &dv1.UploadFileRequest{
-				Data: &dv1.UploadFileRequest_Info{
-					Info: &dv1.FileInfo{
-						FileId:   fileId,
-						FilePath: fileName,
-					},
-				},
-			},
-		},
-	})
-	if err != nil {
-		logger.WithError(err).Error("failed to ready daemon for file upload")
-		return fileId, err
-	}
-	logger.Info("waiting for done signal")
-	<-done
-	if listenerErr != nil {
-		logger.WithError(listenerErr).Error("failed to ready daemon for file upload")
-		return fileId, listenerErr
-	}
-	logger.Info("daemon ready for file upload")
-
-	// chunk file and upload
-	err = c.streamFile(ctx, logger, buf, fileId)
-	if err != nil {
-		logger.WithError(err).Error("failed to upload chunked file")
-		return fileId, err
-	}
-
-	return fileId, nil
-}
-
-// OS
-
-func (c *controller) CheckForOSUpdates(ctx context.Context, logger chassis.Logger) (*v1.CheckForSystemUpdatesResponse, error) {
-	if !c.systemUpdateLock.TryLock() {
-		logger.Warn("call to check for system updates while another check is already in progress")
-		return nil, fmt.Errorf("system update check already in progress")
-	}
-	defer c.systemUpdateLock.Unlock()
-
-	var (
-		response = &v1.CheckForSystemUpdatesResponse{}
-	)
-
-	// get the os update diff from the daemon
-	done := make(chan bool)
-	go async.RegisterListener(ctx, c.broadcaster, &async.ListenerOptions[*dv1.OSUpdateDiff]{
-		Callback: func(event *dv1.OSUpdateDiff) (bool, error) {
-			response.OsDiff = event.Description
-			done <- true
-			return true, nil
-		},
-	}).Listen(ctx)
-	err := com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_RequestOsUpdateDiff{},
-	})
-	if err != nil {
-		return nil, err
-	}
-	<-done
-
-	// get the current daemon version from the daemon
-	done = make(chan bool)
-	go async.RegisterListener(ctx, c.broadcaster, &async.ListenerOptions[*dv1.CurrentDaemonVersion]{
-		Callback: func(event *dv1.CurrentDaemonVersion) (bool, error) {
-			response.DaemonVersions = &v1.DaemonVersions{
-				Current: &v1.DaemonVersion{
-					Version:    event.Version,
-					VendorHash: event.VendorHash,
-					SrcHash:    event.SrcHash,
-				},
-			}
-			done <- true
-			return true, nil
-		},
-	}).Listen(ctx)
-	err = com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_RequestCurrentDaemonVersion{},
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	// get latest available daemon version
-	latest, err := getLatestDaemonVersion()
-	if err != nil {
-		return nil, err
-	}
-	response.DaemonVersions.Latest = latest
-
-	return response, nil
-}
-
-func (c *controller) AutoUpdateOS(logger chassis.Logger) {
-	cr := cron.New()
-	f := func() {
-		ctx := context.Background()
-		err := c.UpdateOS(ctx, logger)
-		if err != nil {
-			logger.WithError(err).Error("failed to run auto os update job")
-		}
-	}
-	cron := chassis.GetConfig().GetString(osAutoUpdateCronConfigKey)
-	logger.WithField("cron", cron).Info("setting os auto-update interval")
-	_, err := cr.AddFunc(cron, f)
-	if err != nil {
-		logger.WithError(err).Panic("failed to initialize auto-update for os")
-	}
-	cr.Start()
-}
-
-func (u *controller) UpdateOS(ctx context.Context, logger chassis.Logger) error {
-	logger.Info("updating os")
-	settings := &v1.DeviceSettings{}
-	err := kvclient.Get(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, settings)
-	if err != nil {
-		logger.WithError(err).Error("failed to get device settings")
-		return err
-	}
-
-	if !settings.AutoUpdateOs {
-		logger.Info("auto update sytem not enabled")
-		return nil
-	}
-
-	updates, err := u.CheckForOSUpdates(ctx, logger)
-	if err != nil {
-		logger.WithError(err).Error("failed to check for system updates")
-		return err
-	}
-
-	// if the daemon needs updating, install it along with the os updates
-	// otherwise just install the os update
-	if updates.DaemonVersions.Current != updates.DaemonVersions.Latest {
-		err = com.Send(&dv1.ServerMessage{
-			Message: &dv1.ServerMessage_ChangeDaemonVersionCommand{
-				ChangeDaemonVersionCommand: &dv1.ChangeDaemonVersionCommand{
-					Version:    updates.DaemonVersions.Latest.Version,
-					VendorHash: updates.DaemonVersions.Latest.VendorHash,
-					SrcHash:    updates.DaemonVersions.Latest.SrcHash,
-				},
-			},
-		})
-	} else {
-		err = com.Send(&dv1.ServerMessage{
-			Message: &dv1.ServerMessage_InstallOsUpdateCommand{},
-		})
-	}
-	if err != nil {
-		logger.WithError(err).Error("failed to install system update")
-		return err
-	}
-
-	return nil
-}
-
-func (c *controller) EnableWireguard(ctx context.Context, logger chassis.Logger) error {
-	var (
-		err    error
-		config = &dv1.WireguardConfig{}
-	)
-
-	// check if wireguard is already enabled
-	result, err := kvclient.List(ctx, config)
-	if err != nil {
-		return err
-	}
-	if result != nil {
-		logger.Warn("the Wireguard server is already enabled")
-		return nil
-	}
-
-	key, err := wgtypes.GeneratePrivateKey()
-	if err != nil {
-		return err
-	}
-
-	// create wireguard config
-	config = &dv1.WireguardConfig{
-		Interfaces: []*dv1.WireguardInterface{
-			{
-				Id:         uuid.New().String(),
-				Name:       DefaultWireguardInterface,
-				PrivateKey: key.String(),
-				Ips: []string{
-					"10.100.0.1/24",
-				},
-				ListenPort: 51820,
-				Peers:      []*dv1.WireguardPeer{},
-			},
-		},
-	}
-
-	// command daemon to initialize
-	done := make(chan bool)
-	var listenerErr error
-	go func() {
-		listenerErr = async.RegisterListener(ctx, c.broadcaster, &async.ListenerOptions[*dv1.WireguardInterfaceAdded]{
-			Callback: func(event *dv1.WireguardInterfaceAdded) (bool, error) {
-				done <- true
-				if event.Error != nil {
-					return true, fmt.Errorf(event.Error.Error)
-				}
-				return true, nil
-			},
-			Timeout: 30 * time.Second,
-		}).Listen(ctx)
-		if listenerErr != nil {
-			done <- true
-		}
-	}()
-	err = com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_AddWireguardInterface{
-			AddWireguardInterface: &dv1.AddWireguardInterface{
-				// TODO: handle this better
-				Interface: config.Interfaces[0],
-			},
-		},
-	})
-	if err != nil {
-		return err
-	}
-	<-done
-	if listenerErr != nil {
-		return listenerErr
-	}
-
-	// save config to blueprint
-	_, err = kvclient.Set(ctx, kvclient.WIREGUARD_CONFIG_KEY, config)
-	if err != nil {
-		return err
-	}
-
-	// enable feature in blueprint
-	settings := &v1.DeviceSettings{}
-	err = kvclient.Get(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, settings)
-	if err != nil {
-		return err
-	}
-	settings.LocatorSettings = &v1.LocatorSettings{
-		Enabled:  true,
-		Locators: make([]*v1.Locator, 0),
-	}
-	_, err = kvclient.Set(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, settings)
-	if err != nil {
-		return fmt.Errorf("failed to save settings")
-	}
-
-	return nil
-}
-
-func (c *controller) DisableWireguard(ctx context.Context, logger chassis.Logger) error {
-	var (
-		err error
-	)
-
-	// disable on daemon
-	done := make(chan bool)
-	var listenerErr error
-	go func() {
-		listenerErr = async.RegisterListener(ctx, c.broadcaster, &async.ListenerOptions[*dv1.WireguardInterfaceRemoved]{
-			Callback: func(event *dv1.WireguardInterfaceRemoved) (bool, error) {
-				done <- true
-				if event.Error != nil {
-					return true, fmt.Errorf(event.Error.Error)
-				}
-				return true, nil
-			},
-			Timeout: 30 * time.Second,
-		}).Listen(ctx)
-		if listenerErr != nil {
-			done <- true
-		}
-	}()
-	err = com.Send(&dv1.ServerMessage{
-		Message: &dv1.ServerMessage_RemoveWireguardInterface{
-			RemoveWireguardInterface: &dv1.RemoveWireguardInterface{
-				Name: DefaultWireguardInterface,
-			},
-		},
-	})
-	if err != nil {
-		return err
-	}
-	<-done
-	if listenerErr != nil {
-		return listenerErr
-	}
-
-	// delete config from blueprint
-	_, err = kvclient.Delete(ctx, kvclient.WIREGUARD_CONFIG_KEY, &dv1.WireguardConfig{})
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// CONTAINERS
-
-func (c *controller) CheckForContainerUpdates(ctx context.Context, logger chassis.Logger) ([]*v1.ImageVersion, error) {
-	var (
-		images []*v1.ImageVersion
-	)
-
-	// populate current versions (from k8s)
-	images, err := c.k8sclient.CurrentImages(ctx)
-	if err != nil {
-		logger.WithError(err).Error("failed to get current container versions")
-		return nil, err
-	}
-
-	// populate latest versions (from registry)
-	images, err = getLatestImageTags(ctx, images)
-	if err != nil {
-		logger.WithError(err).Error("failed to get latest image versions")
-		return nil, err
-	}
-
-	return images, err
-}
-
-func (c *controller) AutoUpdateContainers(logger chassis.Logger) {
-	cr := cron.New()
-	f := func() {
-		ctx := context.Background()
-		err := c.UpdateContainers(ctx, logger)
-		if err != nil {
-			logger.WithError(err).Error("failed to run auto container update job")
-		}
-	}
-	cron := chassis.GetConfig().GetString(containerAutoUpdateCronConfigKey)
-	logger.WithField("cron", cron).Info("setting container auto-update interval")
-	_, err := cr.AddFunc(cron, f)
-	if err != nil {
-		logger.WithError(err).Panic("failed to initialize auto-update for system containers")
-	}
-	cr.Start()
-}
-
-func (c *controller) UpdateContainers(ctx context.Context, logger chassis.Logger) error {
-	logger.Info("updating containers")
-	settings := &v1.DeviceSettings{}
-	err := kvclient.Get(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, settings)
-	if err != nil {
-		logger.WithError(err).Error("failed to get device settings")
-		return err
-	}
-
-	// TODO: should this be a different setting?
-	if !settings.AutoUpdateOs {
-		logger.Info("auto update sytem not enabled")
-		return nil
-	}
-
-	images, err := c.CheckForContainerUpdates(ctx, logger)
-	if err != nil {
-		logger.WithError(err).Error("failed to check for system container updates")
-		return err
-	}
-
-	for _, image := range images {
-		log := logger.WithFields(chassis.Fields{
-			"image":           image.Image,
-			"current_version": image.Current,
-			"latest_version":  image.Latest,
-		})
-		if semver.Compare(image.Latest, image.Current) == 1 {
-			log.Info("updating image")
-			err := com.Send(&dv1.ServerMessage{
-				Message: &dv1.ServerMessage_SetSystemImageCommand{
-					SetSystemImageCommand: &dv1.SetSystemImageCommand{
-						CurrentImage:   fmt.Sprintf("%s:%s", image.Image, image.Current),
-						RequestedImage: fmt.Sprintf("%s:%s", image.Image, image.Latest),
-					},
-				},
-			})
-			if err != nil {
-				log.WithError(err).Error("failed to update system container image")
-				// don't return, try to update other containers
-			}
-			// TODO: this is a hack, should really be event-driven
-			time.Sleep(3 * time.Second)
-		} else {
-			log.Info("no update needed")
-		}
-	}
-
-	return nil
-}
-
-// DEVICE
-
-func (c *controller) GetServerSettings(ctx context.Context) (*v1.DeviceSettings, error) {
-	settings := &v1.DeviceSettings{}
-	err := kvclient.Get(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, settings)
-	if err != nil {
-		return nil, err
-	}
-
-	settings.AdminUser.Password = "" // don't return the password
-
-	return settings, nil
-}
-
-func (c *controller) SetServerSettings(ctx context.Context, logger chassis.Logger, settings *v1.DeviceSettings) error {
-
-	// set the device settings on the host (via the daemon)
-	err := c.saveSettings(ctx, logger, &dv1.SaveSettingsCommand{
-		AdminPassword:  settings.AdminUser.Password,
-		TimeZone:       settings.Timezone,
-		EnableSsh:      settings.EnableSsh,
-		TrustedSshKeys: settings.TrustedSshKeys,
-	})
-	if err != nil {
-		return err
-	}
-
-	// salt and hash given password if set
-	// otherwise, get existing value from cache
-	if settings.AdminUser.Password != "" {
-		logger.Info("updating admin user password")
-		// get seed salt value from blueprint
-		seed, err := getSaltValue(ctx)
-		if err != nil {
-			return err
-		} else {
-			// salt & hash the meat before you put in on the grill
-			settings.AdminUser.Password = hashPassword(settings.AdminUser.Password, []byte(seed))
-		}
-	} else {
-		existingSettings := &v1.DeviceSettings{}
-		err := kvclient.Get(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, existingSettings)
-		if err != nil {
-			return fmt.Errorf(ErrFailedToGetSettings)
-		}
-		settings.AdminUser.Password = existingSettings.AdminUser.Password
-	}
-
-	_, err = kvclient.Set(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, settings)
-	if err != nil {
-		return errors.New(ErrFailedToCreateSettings)
-	}
-
-	return nil
-}
-
-func (c *controller) IsDeviceSetup(ctx context.Context) (bool, error) {
-	// list is used to get all the `DeviceSettings` objects in the key-value store
-	// it will not fail if the key does not exist like `Get` would
-	settings := &v1.DeviceSettings{}
-	list, err := kvclient.List(ctx, settings)
-	if err != nil {
-		return false, errors.New(ErrFailedToGetSettings)
-	}
-
-	if len(list) == 0 {
-		return false, nil
-	}
-	return true, nil
-}
-
-func (c *controller) InitializeDevice(ctx context.Context, logger chassis.Logger, settings *v1.DeviceSettings) error {
-	yes, err := c.IsDeviceSetup(ctx)
-	if err != nil {
-		return errors.New(ErrFailedToGetSettings)
-	} else if yes {
-		logger.Warn("device is already set up")
-		return errors.New(ErrDeviceAlreadySetup)
-	}
-
-	// set the device settings on the host (via the daemon)
-	err = c.saveSettings(ctx, logger, &dv1.SaveSettingsCommand{
-		AdminPassword:  settings.AdminUser.Password,
-		TimeZone:       settings.Timezone,
-		EnableSsh:      settings.EnableSsh,
-		TrustedSshKeys: settings.TrustedSshKeys,
-	})
-	if err != nil {
-		return err
-	}
-
-	// get seed salt value from blueprint
-	seed, err := getSaltValue(ctx)
-	if err != nil {
-		return err
-	} else {
-		// salt & hash the meat before you put in on the grill
-		settings.AdminUser.Password = hashPassword(settings.AdminUser.Password, []byte(seed))
-	}
-
-	_, err = kvclient.Set(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, settings)
-	if err != nil {
-		return errors.New(ErrFailedToCreateSettings)
-	}
-
-	return nil
-}
-
-func (c *controller) Login(ctx context.Context, username, password string) (string, error) {
-	settings := &v1.DeviceSettings{}
-	err := kvclient.Get(ctx, kvclient.DEFAULT_DEVICE_SETTINGS_KEY, settings)
-	if err != nil {
-		return "", errors.New(ErrFailedToGetSettings)
-	}
-
-	salt, err := getSaltValue(ctx)
-	if err != nil {
-		return "", errors.New(ErrFailedToGetSettings)
-	}
-
-	// Check if the password is correct. If not, return an error
-	if hashPassword(password, []byte(salt)) != settings.AdminUser.Password {
-		return "", errors.New("invalid username or password")
-	}
-
-	// TODO: forge token
-
-	return "JWT_TOKEN", nil
-}
 
 // helper functions
